% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_geometric.R
\name{seq_geometric}
\alias{seq_geometric}
\title{Generate Geometric Sequence}
\usage{
seq_geometric(from, to, by.ratio, length.out = NULL)
}
\arguments{
\item{from}{Number. First term.  Scale factor, equal to the sequence's start value.}

\item{to}{Number. If the value closest to the last term in the finite series you want is not 'to' in the identity series, the value closest to 'to' and closest to 'from' in the identity series will be the last term.
If 'to' is contained on an geometric sequence, it is the last term.}

\item{by.ratio}{Number. Common ratio.}

\item{length.out}{Number. Desired length of the sequence. A non-negative number. It will be rounded up if fractional.}
}
\value{
numeric arguments return a vector of type integer or double: programmers should not rely on which.
}
\description{
Function to generate a sequence of equal ratios, also known as a geometric sequence.
Geometric sequence, also known as a Geometric progression, is a sequence of non-zero numbers where each term after the first is found by multiplying the previous one by a fixed, non-zero number called the common ratio.
This function can only create finite sequences of numbers, not infinite sequences.
}
\details{
Numerical inputs should all be finite (that is, not infinite, NaN or NA).
}
\examples{
seq_geometric(1, 10, 2)
seq_geometric(1, 10, 2, length.out = 3)
seq_geometric(1, 0.001, 0.1)
seq_geometric(1, 0.001, 0.1, length.out = 3)
}
